refactor files so each class has its own file organize imports
