+++ pyfritzhome/__init__.py
@@ -1,5 +1,6 @@
 from .errors import InvalidError, LoginError
-from .fritzhome import Fritzhome, FritzhomeDevice
+from .fritzhome import Fritzhome
+from .fritzhomedevice import FritzhomeDevice
 
 __all__ = [
     "LoginError",
